src/requirements.txt
aniso8601==1.0.0
blinker==1.4
clay-flask==2.1.5
Flask==0.10.1
Flask-Admin==1.2.0
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-RESTful==0.3.4
Flask-SQLAlchemy==2.0
Flask-User==0.6.6
Flask-WTF==0.12
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
numpy==1.9.2
passlib==1.6.5
pdfminer==20110515
pdftables==0.0.4
py-bcrypt==0.4
pycrypto==2.6.1
pymongo==3.0.3
pytz==2015.4
redis==2.10.3
requests==2.7.0
six==1.9.0
SQLAlchemy==1.0.8
uWSGI==2.0.11.1
Werkzeug==0.10.4
wheel==0.24.0
WTForms==2.0.2
